backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.8.30
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.3
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-import-export==4.3.3
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-mathfilters==1.0.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
import-export==0.3.1
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
namex==0.0.8
numpy==2.1.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2024.2
RapidFuzz==3.13.0
redis==5.2.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==77.0.1
shortuuid==1.0.13
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
syncedlyrics==1.0.1
tablib==3.7.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.29.2
virtualenvwrapper-win==1.2.7
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
